+++ reactive/reactive_node_test.ts
@@ -1,0 +1,10 @@
+// Copyright the Deft+ authors. All rights reserved. Apache-2.0 license
+
+import { describe, test } from '@std/testing/bdd';
+import { expect } from '@std/expect';
+
+describe('reactive / ReactiveNode()', () => {
+  test('should pass', () => {
+    expect(true).toBe(true);
+  });
+});
