+++ scripts/_ascii_table.ts
@@ -1,3 +1,4 @@
+// Copyright the Deft+ authors. All rights reserved. Apache-2.0 license
 // deno-lint-ignore-file
 
 export interface AsciiTableOptions {
