+++ scripts/version.ts
@@ -1,0 +1,3 @@
+// Copyright the Deft+ authors. All rights reserved. Apache-2.0 license
+
+// TODO(@miguelbogota): Implement versioning here.
