--- store/mod.ts
@@ -1,9 +1,0 @@
-/**
- * @license
- * Copyright Deft+ All Rights Reserved.
- *
- * Use of this source code is governed by an Apache-2.0 license that can be
- * found in the LICENSE file at https://github.com/deft-plus/fragment.js/blob/latest/LICENCE
- */
-
-export * from '@/store/store.ts';
